rename config and release swin baselines
