+++ projects/dino/configs/models/dino_swin_base_384.py
